#1 - Rename Types\Int to Types\Integer

+++ tests/Conditions/Binaries/AndConditionTest.php
@@ -19,7 +19,7 @@ class AndConditionTest extends PHPUnit_Framework_TestCase
         $condition = (new Conditions\Equal('name', new Types\String('rainbow')))
             ->and(
                 (new Conditions\Equal('taste', new Types\String('burger')))
-                ->or(new Conditions\Equal('rank', new Types\Int(42)))
+                ->or(new Conditions\Equal('rank', new Types\Integer(42)))
             );
 
         $this->assertSame($condition->toString($this->escaper), "name = 'rainbow' AND (taste = 'burger' OR rank = 42)");
@@ -39,7 +39,7 @@ class AndConditionTest extends PHPUnit_Framework_TestCase
     {
         $conditionA = new Conditions\Equal('name', new Types\String('rainbow'));
         $conditionB = new Conditions\Equal('taste', new Types\String('burger'));
-        $conditionC = new Conditions\Equal('rank', new Types\Int('42'));
+        $conditionC = new Conditions\Equal('rank', new Types\Integer('42'));
         $conditionD = new Conditions\Equal('author', new Types\String('julian'));
 
         $orComposite1  = new Conditions\Binaries\OrCondition($conditionA, $conditionB);
